Grouped dataclasses
